# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Evidencija Recepata API",
        "description": "REST API za upravljanje receptima, sastojcima i korisnicima.",
        "contact": {
            "email": "admin@evidencijarecepata.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Glavni server"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Registracija novog korisnika",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ime",
                                    "email",
                                    "lozinka",
                                    "tip"
                                ],
                                "properties": {
                                    "ime": {
                                        "type": "string",
                                        "example": "Marko Marković"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "marko@email.com"
                                    },
                                    "lozinka": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "tajna123"
                                    },
                                    "tip": {
                                        "type": "string",
                                        "enum": [
                                            "amater",
                                            "profesionalac",
                                            "dobavljac"
                                        ],
                                        "example": "amater"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Uspešna registracija",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Uspešna registracija"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc123..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Prijava korisnika",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "lozinka"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "marko@email.com"
                                    },
                                    "lozinka": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "tajna123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uspešna prijava",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Uspešna prijava"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc123..."
                                        },
                                        "korisnik": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Pogrešni kredencijali"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Odjava korisnika",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Uspešna odjava",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Uspešno ste se izlogovali"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautorizovan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Unesi token dobijen pri registraciji/prijavi",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}